enclosed voxel level generator
The random terrain function to select a position for a tile to be setup is coming from a youtuber. The chunk vertices/triangles/indexes is coming from Craig Perko's old minecraft tutorial. The walls texture, i have figured out how to do it with help from stackoverflow and the unity3d forums.

thank you for reading me.
steve chassé